Move hivemetastore topology back as non plugin
