+++ cmd/punch/generate.go
@@ -23,6 +23,7 @@ import (
 	"os"
 
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/eks"
+	_ "github.com/datapunchorg/punch/pkg/topologyimpl/hivemetastore"
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/kafkabridge"
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/kafkaonmsk"
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/kyuubioneks"
