+++ pkg/topologyimpl/hivemetastore/topologyhandler.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package hivemetastore
 
 import (
 	"fmt"
@@ -31,8 +31,6 @@ import (
 
 type TopologyHandler struct {
 }
-
-var Handler TopologyHandler
 
 func (t *TopologyHandler) Generate() (framework.Topology, error) {
 	topology := GenerateHiveMetastoreTopology()
